Add guest count and room status properties
- Added `NumberOfGuest` property to `Reservation` class.
- Updated `TotalPrice` calculation to explicitly cast to `decimal`.
- Changed `PricePerNight` type from `decimal` to `double` in `Room` class.
- Introduced `Status` property in `Room` class, defaulting to `RoomStatus.Available`.
- Initialized `Reservations` property as an empty list.

+++ SE491_Project/Data/Entities/Room.cs
@@ -8,7 +8,8 @@
 
         public string Type { get; set; }        
 
-        public decimal PricePerNight { get; set; }          
+        public double PricePerNight { get; set; }  
+       
 
         public RoomStatus Status { get; set; } = RoomStatus.Available;
